ff-userjs: Stop including nl-BE
Having both english and dutch in the accepted languages probably makes
for easier fingerprinting. I have never come across a site that had no
english variant, did have a dutch one but not as a default. So the use
seems dubious. Would be different if english was only a fallback
language.