nixos: Add libPurple-Matrix to Bitlbee
The Nix community has moved to Matrix so I need this to be part of it.
It lacks encryption support and rendering of answers is somewhat
unpleasant. Adding Pantalaimon on top for encryption sounds like the
easiest way forward but I've already joined so many channels that the
initial sync is bigger than the plugin supports.